+++ changeLog.md
@@ -1,5 +1,36 @@
 # Change log #
 
+## v0.4.0
+
+- added a schema manager filter that suggests attribute values
+  according to labelled entries from plugins in text mode
+- added an attribute condition in the config file
+
+## v0.3.2
+
+- send debugging message to stderr and to an exception window when no
+  labelled items are found for the action. It prints the used
+  providers and the arguments, they are configured with
+- plugin config
+  - moved default config file from `samples` to `config` folder
+  - moved XPath for getting `<prefixDef>` from arguments section to
+    conditions section, because it's not an argument of the plugin,
+    but evaluated by the editor
+- removed the matching of @ident of `<prefixDef>` from the activation
+  XPath in the author mode action, because these made to much
+  assumptions
+
+## v0.3.1
+
+- made selection dialogue configurable through author mode action xml
+- changed API for selection dialogs to handle multiple selection
+- reflected these changes in the wrappers around the oxygen dialog and
+  the ediarum dialog
+- discovered bugs in these dialogs and documented them in the wrappers
+- moved reusable code to abstract base class AbstractOperation
+  - saw some problems with protected fields of this base class
+  - Should abstract base classed be used to share state?
+
 ## v0.3.0
 
 - generate selections from `<prefixDef>` through configurable Java